refactor(react): ♻️ ReactDOM вынесен в отдельный файл
